+++ backend/src/common/server-application/server-application.ts
@@ -1,5 +1,5 @@
 import { authController } from '~/bundles/auth/auth.js';
-import { resumeScoreController } from '~/bundles/resume-score/resume-score.js';
+import { resumeController } from '~/bundles/resumes/resumes.js';
 import { templateController } from '~/bundles/templates/templates.js';
 import { userController } from '~/bundles/users/users.js';
 import { config } from '~/common/config/config.js';
@@ -14,7 +14,7 @@ const apiV1 = new ServerAppApi(
     config,
     ...authController.routes,
     ...userController.routes,
-    ...resumeScoreController.routes,
+    ...resumeController.routes,
     ...templateController.routes,
 );
 const serverApp = new ServerApp({
